Update UI layout and details in weather app
This commit modifies details displayed in the Weather App's user interface. The changes include updating the placement of icons for temperature and wind in the html file and adding information about the country in the app.js file. The goal is to enhance user experience by providing more precise location data and improving the layout of weather cards.